+++ frontend/src/pages/Admin/Relatorios.jsx
@@ -1,0 +1,9 @@
+function RelatoriosLogin() {
+  return (
+  <main className="font-viminalis py-4">
+    <p>Relatorios Admin</p>
+  </main>
+  )
+}
+
+export default RelatoriosLogin