+++ frontend/src/pages/Admin/Vendas.jsx
@@ -1,0 +1,9 @@
+function AdminVendas() {
+  return (
+  <main className="font-viminalis py-4">
+    <p>Vendas Admin</p>
+  </main>
+  )
+}
+
+export default AdminVendas